Move generate commands to new IaCGenerator class
