--- dbt_platform_helper/domain/environment.py
@@ -1,3 +1,0 @@
-class Environment:
-    def __init__(self, env_name):
-        self.env_name = env_name
